--- CombineCoreData/README.md
@@ -1,3 +1,0 @@
-# CombineCoreData
-
-A description of this package.
