--- CombineCoreData/Sources/CombineCoreData/CoreDataStore.swift
@@ -1,90 +1,0 @@
-import CoreData
-
-enum StorageType {
-    case persistent, inMemory
-}
-
-extension NSManagedObject {
-    class var entityName: String {
-        return String(describing: self).components(separatedBy: ".").last!
-    }
-}
-
-protocol EntityCreating {
-    var viewContext: NSManagedObjectContext { get }
-    func createEntity<T: NSManagedObject>() -> T
-}
-
-extension EntityCreating {
-    func createEntity<T: NSManagedObject>() -> T {
-        T(context: viewContext)
-    }
-}
-
-protocol CoreDataFetchResultsPublishing {
-    var viewContext: NSManagedObjectContext { get }
-    func publisher<T: NSManagedObject>(fetch request: NSFetchRequest<T>) -> CoreDataFetchResultsPublisher<T>
-}
-
-extension CoreDataFetchResultsPublishing {
-    func publisher<T: NSManagedObject>(fetch request: NSFetchRequest<T>) -> CoreDataFetchResultsPublisher<T> {
-        return CoreDataFetchResultsPublisher(request: request, context: viewContext)
-    }
-}
-
-protocol CoreDataDeleteModelPublishing {
-    var viewContext: NSManagedObjectContext { get }
-    func publisher(delete request: NSFetchRequest<NSFetchRequestResult>) -> CoreDataDeleteModelPublisher
-}
-
-extension CoreDataDeleteModelPublishing {
-    func publisher(delete request: NSFetchRequest<NSFetchRequestResult>) -> CoreDataDeleteModelPublisher {
-        return CoreDataDeleteModelPublisher(delete: request, context: viewContext)
-    }
-}
-
-protocol CoreDataSaveModelPublishing {
-    var viewContext: NSManagedObjectContext { get }
-    func publisher(save action: @escaping Action) -> CoreDataSaveModelPublisher
-}
-
-extension CoreDataSaveModelPublishing {
-    func publisher(save action: @escaping Action) -> CoreDataSaveModelPublisher {
-        return CoreDataSaveModelPublisher(action: action, context: viewContext)
-    }
-}
-
-protocol CoreDataStoring: EntityCreating, CoreDataFetchResultsPublishing, CoreDataDeleteModelPublishing, CoreDataSaveModelPublishing {
-    var viewContext: NSManagedObjectContext { get }
-}
-
-class CoreDataStore: CoreDataStoring {
-    
-    private let container: NSPersistentContainer
-    
-    static var `default`: CoreDataStoring = {
-        return CoreDataStore(name: "GitignoreData", in: .persistent)
-    }()
-    
-    var viewContext: NSManagedObjectContext {
-        return self.container.viewContext
-    }
-    
-    init(name: String, in storageType: StorageType) {
-        self.container = NSPersistentContainer(name: name)
-        self.setupIfMemoryStorage(storageType)
-        self.container.loadPersistentStores { (storeDescription, error) in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        }
-    }
-    
-    private func setupIfMemoryStorage(_ storageType: StorageType) {
-        if storageType  == .inMemory {
-            let description = NSPersistentStoreDescription()
-            description.url = URL(fileURLWithPath: "/dev/null")
-            self.container.persistentStoreDescriptions = [description]
-        }
-    }
-}
