+++ Effects/Package.swift
@@ -11,24 +11,15 @@ let package = Package(
         .library(
             name: "Effects",
             targets: ["Effects"]),
-        .library(
-            name: "Local",
-            targets: ["Local"]),
     ],
     dependencies: [
         .package(url: "https://github.com/pointfreeco/swift-composable-architecture", branch: "main"),
-        .package(path: "../CombineCoreData"),
     ],
     targets: [
         .target(
             name: "Effects",
             dependencies: [
                 .product(name: "ComposableArchitecture", package: "swift-composable-architecture")
-            ]),
-        .target(
-            name: "Local",
-            dependencies: [
-                .product(name: "CombineCoreData", package: "CombineCoreData")
             ]),
     ]
 )
