--- Effects/Sources/Local/Common /BaseCoreDataService.swift
@@ -1,7 +1,0 @@
-public class BaseCoreDataService  {
-    unowned let data : ServiceDataType
-    
-    init(data : ServiceDataType){
-        self.data = data
-    }
-}
