--- Effects/Sources/Local/Model/GitignoreList.xcdatamodeld/Model.xcdatamodel/contents
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="20086" systemVersion="21F79" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="GitignoreList" representedClassName="Local.GitignoreList" syncable="YES" codeGenerationType="class">
-        <attribute name="date" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="gitignoreString" optional="YES" attributeType="String"/>
-        <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="tags" optional="YES" attributeType="Transformable"/>
-        <attribute name="title" optional="YES" attributeType="String"/>
-    </entity>
-    <elements>
-        <element name="GitignoreList" positionX="-63" positionY="-18" width="128" height="104"/>
-    </elements>
-</model>
