+++ gitignore/Sources/UI/Main/MainView.swift
@@ -27,25 +27,8 @@ struct MainView: View {
     }
     
     var body: some View {
-        Color.backgroundColor
-            .ignoresSafeArea()
-            .padding(.leading,-10)
-        
         WithViewStore(self.store.scope(state: ViewState.init)) { viewStore in
             VStack{
-                HStack{
-                    Spacer()
-                    Button(action: {viewStore.send(.addSheetButtonDidTap)}, label: {
-                        Image(systemName: "plus")
-                            .font(.title2)
-                    })
-                    .sheet(isPresented: viewStore.binding(
-                        get: \.addSheetStatus,send: MainAction.addSheetButtonDidTap)){
-                            SheetView(store: store)
-                    }
-                    .buttonStyle(ToolBarButtonStyle())
-                }
-                .padding()
                 Spacer()
                 VStack{
                     TitleLogo()
